requirements.txt
psycopg2-binary==2.9.9
pgvector==0.2.4
numpy==1.24.3
gdown==4.7.1
crewai
openai
psycopg2-binary
pgvector
python-dotenv
numpy